% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_required_args}
\alias{check_required_args}
\title{Checked required arg}
\usage{
check_required_args(required_arg, passed)
}
\arguments{
\item{required_arg}{Vector of required arguments.}

\item{passed}{Vector of passed arguments.}
}
\description{
Checked required arg
}
