% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloads.R
\name{download_file}
\alias{download_file}
\title{download hyperlink}
\usage{
download_file(conn, dest_path, overwrite = FALSE, ...)
}
\arguments{
\item{conn}{Url.}

\item{dest_path}{Folder where to save files.}

\item{overwrite}{If TRUE overwrite the file.}

\item{...}{Any arguments of `download.file`.}
}
\description{
Use download.file to download from hyperlink
}
