% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloader-package.r
\docType{package}
\name{downloader}
\alias{downloader-package}
\alias{downloader}
\title{downloader: a package for making it easier to download files over https}
\description{
This package provides a wrapper for the download.file function,
making it possible to download files over https on Windows, Mac OS X, and
other Unix-like platforms. The RCurl package provides this functionality
(and much more) but can be difficult to install because it must be compiled
with external dependencies. This package has no external dependencies, so
it is much easier to install.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/wch/downloader}
  \item Report bugs at \url{https://github.com/wch/downloader/issues}
}

}
\author{
\strong{Maintainer}: Winston Chang \email{winston@stdout.org}

}
