% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{extract_hyperlinks}
\alias{extract_hyperlinks}
\title{extract_hypelinks}
\usage{
extract_hyperlinks(path_file)
}
\arguments{
\item{path_file}{Path to xlsx file with hyperlinks}
}
\value{
A vector of hyperlinks.
}
\description{
Use XML to extract hyperlinks and generate a list
}
