% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloads.R
\name{retry_download_}
\alias{retry_download_}
\title{retry download internal}
\usage{
retry_download_(conn, dest_path, max_attempts, sleep_time, ...)
}
\arguments{
\item{conn}{Url.}

\item{dest_path}{Folder where to save files.}

\item{max_attempts}{How many times to try to download.}

\item{sleep_time}{How many second to wait before retring.}

\item{...}{Any `download.file` argument.}
}
\description{
Use download_files, try to download from links and save stout or err in dataframe.
}
