% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloads.R
\name{retry_download}
\alias{retry_download}
\title{retry download}
\usage{
retry_download(conn, dest_path, max_attempts = 1, sleep_time = 0, ...)
}
\arguments{
\item{conn}{Vectors of connections (urls).}

\item{dest_path}{Folder where to save files.}

\item{max_attempts}{How many times to try to download.}

\item{sleep_time}{How many second to wait before retrying.}

\item{...}{Any arguments of `download.file`.}

\item{timeout_user}{Seconds after witch the download will be stopped. Default is 3600 sec.}
}
\description{
Use download_files, try to download from links and save stout or err in dataframe.
}
